[iOS] Fabric: Fixes textinput not respond when maxLength > 5.5 quadrillion

+++ packages/react-native/ReactCommon/react/renderer/components/textinput/BaseTextInputProps.h
@@ -60,7 +60,7 @@ class BaseTextInputProps : public ViewProps, public BaseTextProps {
   // TODO: Rename to `tintColor` and make universal.
   SharedColor underlineColorAndroid{};
 
-  int maxLength{};
+  int64_t maxLength{};
 
   /*
    * "Private" (only used by TextInput.js) props
